Improve portability of the portable version
This detects when the version of the app being run is the portable version
and changes the directories where some things are stored so that they remain
close to the executable file itself. This makes the app more truly portable
when it is run, e.g. from a thumb drive.

In particular, settings and configuration files, etc. will be stored in a
"Data" folder in the same directory with the executable, and image thumbnails
will be stored in "Other/thumbnails".